+++ apps/ptg-angular-app/src/app/mock/chart.ts
@@ -6,7 +6,7 @@ export const BAR_CHART_3D = {
         data: [107, 120, 635, 203, 300, 203, 300]
       }
     ],
-    categories: ['Africa', 'America', 'Asia', 'Europe', 'Oceania']
+    categories: ['Africa', 'America', 'Asia', 'Europe', 'Oceania', 'Canada', 'Uk']
   }
   
   // Mock data for 2d line chart
