+++ libs/ptg-ui-angular-lib/charts/highcharts/3d/high-3d-bar-chart/high-3d-bar-chart.stories.mdx
@@ -1,0 +1,70 @@
+import { Meta, Story, ArgsTable, Canvas , Preview} from '@storybook/addon-docs';
+
+import * as High3dBarChartComponents from './high-3d-bar-chart.stories.ts'
+
+import High3dBarChartComponent from './high-3d-bar-chart.component';
+
+<Meta title="Component/High3dBarChartComponent/HighBarChartComponentDocs" component={ High3dBarChartComponent }    />
+
+ # High3DBarChart Component
+
+ ---
+ Bar Chart is the most common chart type that is used to compare different categories of data like frequency, count, total, or average displayed in Horizontal bars. 
+
+ We can import the High3DBarChart Component using the following importation : 
+
+```
+    import { High3dBarChartComponent } from ' @ptg-ui/angular'
+```
+
+
+Types of controls supported by High3DBarChart:
+
+ - title
+ - xTitle
+ - yTitle
+
+
+ ## Quick Start
+ 
+ we can create a high3DBarChart using the following tag: 
+
+```
+<ptg-ui-high-3d-bar-chart
+  [data]="data"
+  [remainingOptions]="remainingOptions"
+  title="BAR_CHART_3D"
+  [categories]="categories"
+  xTitle="Categories"
+  yTitle="Values"
+></ptg-ui-high-3d-bar-chart>
+
+```
+
+Following are some controls with their explanation :
+
+ 1] title :
+
+ The title appears at the top of the graph and should describe the graph. 
+
+ This helps the reader immediately know what the graph is. 
+
+
+<Canvas><Story id="component-high3dbarchartcomponent--title" /></Canvas>
+
+
+2] xTitle : 
+
+
+The axis title is a short text displayed alongside the axis. Usually, the axis title shows units of measurement for axis values, but you can put any text in it.
+
+<Canvas><Story id="component-high3dbarchartcomponent--x-title" /></Canvas>
+
+
+
+3] yTitle : 
+
+
+The axis title is a short text displayed alongside the axis. Usually, the axis title shows units of measurement for axis values, but you can put any text in it.
+
+<Canvas><Story id="component-high3dbarchartcomponent--y-title" /></Canvas>
