+++ libs/ptg-ui-angular-lib/charts/highcharts/3d/high-3d-bar-chart/high-3d-bar-chart.stories.ts
@@ -1,0 +1,61 @@
+import { moduleMetadata, Story, Meta } from '@storybook/angular';
+import { High3dBarChartComponent } from './high-3d-bar-chart.component';
+import {ChartModule} from '../../../chart.module';
+import { BAR_CHART_3D, LINE_BAR_DATA, LINE_CHART_2D, MULTI_LINE_CHART_2D, PIE_CHART_3D, STACKED_COLUMN_DATA } from '@ptg-angular-app/mock/chart';
+
+
+export default {
+  title: 'Component/High3dBarChartComponent',
+  component: High3dBarChartComponent,
+  decorators: [
+    moduleMetadata({
+      imports: [ChartModule],
+    }),
+  ],
+} as Meta<High3dBarChartComponent>;
+
+const Template: Story<High3dBarChartComponent> = (
+  args: High3dBarChartComponent
+) => ({
+  props: args,
+});
+
+export const Primary = Template.bind({});
+Primary.args = {
+  data: BAR_CHART_3D.data,
+  categories:BAR_CHART_3D.categories,
+  remainingOptions: {},
+  title: 'BAR_CHART_3D',
+  xTitle: 'Categories',
+  yTitle: 'Values',
+};
+
+export const Title = Template.bind({});
+Title.args = {
+  data: BAR_CHART_3D.data,
+  categories:BAR_CHART_3D.categories,
+  remainingOptions: {},
+  title: 'BAR_CHART_3D',
+  xTitle: 'Categories',
+  yTitle: 'Values',
+};
+
+export const xTitle = Template.bind({});
+xTitle.args = {
+  data: BAR_CHART_3D.data,
+  categories:BAR_CHART_3D.categories,
+  remainingOptions: {},
+  title: 'BAR_CHART_3D',
+  xTitle: 'Categories',
+  yTitle: 'Values',
+};
+
+export const yTitle = Template.bind({});
+yTitle.args = {
+  data: BAR_CHART_3D.data,
+  categories:BAR_CHART_3D.categories,
+  remainingOptions: {},
+  title: 'BAR_CHART_3D',
+  xTitle: 'Categories',
+  yTitle: 'Values',
+};
