+++ libs/ptg-ui-angular-lib/charts/highcharts/3d/high-3d-line-chart/high-3d-line-chart.stories.mdx
@@ -1,0 +1,88 @@
+import { Meta, Story, ArgsTable, Canvas , Preview} from '@storybook/addon-docs';
+
+import * as High3dLineChartComponents from './high-3d-line-chart.stories.ts'
+
+import High3dLineChartComponent from './high-3d-line-chart.component';
+
+<Meta title="Component/High3dLineChartComponent/High3dLineChartComponentDocs" component={ High3dLineChartComponent }    />
+
+
+
+ # High3DLineChart Component
+
+ ---
+
+A line graph is a graphical representation of information that changes over a period of time.
+
+You can plot it by using several points linked by straight lines. It comprises two axes called the “x-axis” and the “y-axis“.
+
+
+We can import the LineChart  using the following importation : 
+```
+    import { High3dLineChartComponent } from ' @ptg-ui/angular'
+
+```    
+
+
+Types of controls supported by High3DLinerChart:
+
+ - title
+ - xTitle
+ - yTitle
+ - chartId
+
+
+ ## Quick Start
+ 
+ we can create a high3DLineChart using the following tag: 
+
+```
+<ptg-ui-high-3d-line-chart
+  [data]="data"
+  [remainingOptions]="remainingOptions"
+  title="LINE_CHART_3D"
+  [categories]="categories"
+  xTitle="Categories"
+  yTitle="Values"
+  chartId="chart-line-3d"
+></ptg-ui-high-3d-line-chart>
+
+```
+
+Following are some controls with their explanation :
+
+ 1] title :
+
+ The title appears at the top of the graph and should describe the graph. 
+
+ This helps the reader immediately know what the graph is. 
+
+
+<Canvas><Story id="component-high3dlinechartcomponent--title" /></Canvas>
+
+
+2] xTitle : 
+
+
+The axis title is a short text displayed alongside the axis. Usually, the axis title shows units of measurement for axis values, but you can put any text in it.
+
+<Canvas><Story id="component-high3dlinechartcomponent--x-title" /></Canvas>
+
+
+
+3] yTitle : 
+
+
+The axis title is a short text displayed alongside the axis. Usually, the axis title shows units of measurement for axis values, but you can put any text in it.
+
+<Canvas><Story id="component-high3dlinechartcomponent--y-title" /></Canvas>
+
+
+4] chartId : 
+
+The chartId attribute specifies a unique id for an chart . The value of the id attribute must be unique.
+
+The id attribute is used to point to a specific style declaration in a style sheet. It is also used by other languages and manipulate the element with the specific id.
+
+
+<Canvas><Story id="component-high3dlinechartcomponent--chart-id" /></Canvas>
