+++ libs/ptg-ui-angular-lib/charts/highcharts/3d/high-3d-pie-chart/high-3d-pie-chart.stories.mdx
@@ -1,0 +1,53 @@
+import { Meta, Story, ArgsTable, Canvas , Preview} from '@storybook/addon-docs';
+
+import * as High3dPieChartComponents from './high-3d-pie-chart.stories.ts'
+
+import High3dPieChartComponent from './high-3d-pie-chart.component';
+
+<Meta title="Component/High3dPieChartComponent/High3dPieChartComponentDocs" component={ High3dPieChartComponent }    />
+
+
+ # High3DPieChart Component
+
+ ---
+
+Pie chart is a type of graph that represents the data in the circular graph.
+
+Pie charts are divided into sections (or 'slices') to represent values of different sizes.
+
+
+We can import the High3DPieChart using the following importation : 
+```
+    import { High3dPieChartComponent } from ' @ptg-ui/angular'
+
+```    
+
+
+Types of controls supported by High3DLinerChart:
+
+ - title
+
+
+ ## Quick Start
+ 
+ we can create a high3DLineChart using the following tag: 
+
+```
+<ptg-ui-high-3d-pie-chart
+  [data]="data"
+  title="PIE_CHART_3D"
+  seriesName=""
+></ptg-ui-high-3d-pie-chart>
+
+```
+
+Following are some controls with their explanation :
+
+ 1] title :
+
+ The title appears at the top of the graph and should describe the graph. 
+
+ This helps the reader immediately know what the graph is. 
+
+
+<Canvas><Story id="component-high3dlinechartcomponent--title" /></Canvas>
